Made the program runnable
Controller: Updated to add actual items to the sale
ReceiptPrinter: Made the printouts look nicer and more in line with a receipt
Sale: Fixed the addition of VAT to add the actual percentage of the item cost.
View: Hardcoded commands from view  which lies in line with the lab instructions.